React CRA 업데이트 문법 검사를 했다 => 임포트 하는 변수와 인터페이스 변수는 이름이 같으면 안된다고 했다 그래서 바꿨다

+++ Client/src/components/Templates/Admin/CategoryChange/CategoryChange.tsx
@@ -1,16 +1,7 @@
 import * as React from 'react'
 import { withRouter, RouteComponentProps } from 'react-router-dom'
 import { CategoryStateInside } from 'store/modules/Category'
-import {
-  Form,
-  InputGroup,
-  Dropdown,
-  DropdownToggle,
-  DropdownItem,
-  DropdownMenu,
-  Input,
-  Button
-} from 'reactstrap'
+import { Form, InputGroup, Dropdown, DropdownToggle, DropdownItem, DropdownMenu, Input, Button } from 'reactstrap'
 import { toast } from 'react-toastify'
 
 import './CategoryChange.css'
@@ -33,13 +24,13 @@ interface Props {
 interface Target {
   target: HTMLInputElement
 }
-interface Dropdown {
+interface Current {
   currentTarget: { textContent: string }
 }
 
 const CategoryChange = withRouter<Props & RouteComponentProps<any>>(
   class CategoryChange extends React.Component<Props & RouteComponentProps<any>> {
-    public changeCategoryInput: HTMLInputElement
+    public changeCategoryInput: any
 
     // 드롭다운 State
     public state = {
@@ -59,7 +50,7 @@ const CategoryChange = withRouter<Props & RouteComponentProps<any>>(
     }
 
     // 변경할 카테고리 선택
-    public handleSelect = (e: Dropdown) => {
+    public handleSelect = (e: Current) => {
       this.props.changeCategoryCategoryChange(e.currentTarget.textContent)
     }
 
@@ -87,35 +78,39 @@ const CategoryChange = withRouter<Props & RouteComponentProps<any>>(
         changeCategoryCategoryValue !== '변경할 카테고리 선택' &&
         loginLogined !== false
       ) {
-        // 카테고리 변경 메소드
-        changeCategory(changeCategoryCategoryValue, changeCategoryInputValue)
-          // 카테고리 변경 성공
-          .then((res: any) => {
-            toast(res.value.data.message)
-            categoryLoad()
-            categoryDone()
-          })
-          // 카테고리 변경 실패
-          .catch((err: any) => {
-            // 사용자의 해킹 시도
-            if (err.response.data.type === 'undefinded token' || err.response.data.type === 'invalid token') {
-              toast('인증된 사용자가 아닙니다 !')
-              // 로그아웃 메소드로 loginLogined false & 세션 스토리지 초기화 & 홈페이지로 이동
-              logout()
-              sessionStorage.clear()
-              history.push('/')
-            }
-            // 사용자의 시도 실패
-            else if (err.response.data.type === 'server error') {
-              toast(err.response.data.message)
-              this.changeCategoryInput.focus()
-              changeCategoryInputChange('')
-              changeCategoryCategoryChange('변경할 카테고리 선택')
-            } else {
-              toast(err.response.data.message)
-              this.changeCategoryInput.focus()
-            }
-          })
+        if (changeCategoryInputValue.toLowerCase() !== 'admin') {
+          // 카테고리 변경 메소드
+          changeCategory(changeCategoryCategoryValue, changeCategoryInputValue)
+            // 카테고리 변경 성공
+            .then((res: any) => {
+              toast(res.value.data.message)
+              categoryLoad()
+              categoryDone()
+            })
+            // 카테고리 변경 실패
+            .catch((err: any) => {
+              // 사용자의 해킹 시도
+              if (err.response.data.type === 'undefinded token' || err.response.data.type === 'invalid token') {
+                toast('인증된 사용자가 아닙니다 !')
+                // 로그아웃 메소드로 loginLogined false & 세션 스토리지 초기화 & 홈페이지로 이동
+                logout()
+                sessionStorage.clear()
+                history.push('/')
+              }
+              // 사용자의 시도 실패
+              else if (err.response.data.type === 'server error') {
+                toast(err.response.data.message)
+                this.changeCategoryInput.focus()
+                changeCategoryInputChange('')
+                changeCategoryCategoryChange('변경할 카테고리 선택')
+              } else {
+                toast(err.response.data.message)
+                this.changeCategoryInput.focus()
+              }
+            })
+        } else {
+          toast('관리자 url로 카테고리 변경이 불가능 합니다 !')
+        }
       } else {
         if (loginLogined === false) {
           toast('관리자만 접근 / 엑세스 가능합니다 !')
